+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release Notes
 All notable changes to this package will be documented in this file.
 
+- Added `color` to **ThumbnailDot** component
+
 # 0.3.39
 - Added `opacity` to **ThumbnailDot** component